create a new file swarm.go and move swarm part code from cluster.go into swarm.go
